Global update

+++ Filigram/main.cpp
@@ -2,7 +2,7 @@
 #include <stdio.h>
 int main()
 {
-	std::srand(static_cast<unsigned int>(std::time(nullptr)));
+	std::srand(std::time(nullptr));
 	MainWindow mWindow;
 
 	mWindow.start();
